Refactor map visualizer
