+++ src/util.ts
@@ -1,5 +1,4 @@
 import { Point, Node, Map, Airplane, Airport } from 'auto-traffic-control'
-import { MapVisualizer } from './map_visualizer'
 
 /**
  *
@@ -120,54 +119,7 @@ export function getAirportForAirplane(map: Map, airplane: Airplane): Airport {
 export function nodeStr(node: Node) {
     return `(${node.getLongitude()}, ${node.getLatitude()})`
 }
-// TODO: function to print out a debug map to see where the pathing is going
+
 export function nodeListStr(nodes: Node[]) {
     return nodes.map(nodeStr).join(' -> ')
-}
-
-export function mapDebugString(map: Map, nodes: Node[]) {
-    const width = map.getWidth()
-    const height = map.getHeight()
-    const airports = map.getAirportsList()
-    const restrictedNodes = map
-        .getRoutingGridList()
-        .filter((node) => node.getRestricted())
-    for (let y = height - 1; y >= 0; y--) {
-        for (let x = 0; x < width; x++) {
-            const index = nodes.findIndex((node) => {
-                const coordinates = getStandardizedCoordinates(
-                    map,
-                    node.getLongitude(),
-                    node.getLatitude(),
-                )
-                return x === coordinates.x && y === coordinates.y
-            })
-            const airportIndex = airports.findIndex((airport) => {
-                const coordinates = getStandardizedCoordinates(
-                    map,
-                    airport.getNode()!.getLongitude(),
-                    airport.getNode()!.getLatitude(),
-                )
-                return x === coordinates.x && y === coordinates.y
-            })
-            const restrictedIndex = restrictedNodes.findIndex((node) => {
-                const coordinates = getStandardizedCoordinates(
-                    map,
-                    node.getLongitude(),
-                    node.getLatitude(),
-                )
-                return x === coordinates.x && y === coordinates.y
-            })
-            if (restrictedIndex > -1) {
-                MapVisualizer.invalid()
-            } else if (airportIndex > -1) {
-                MapVisualizer.airport()
-            } else if (index > -1) {
-                MapVisualizer.path(index)
-            } else {
-                MapVisualizer.empty()
-            }
-        }
-        MapVisualizer.endRow()
-    }
 }
